requirements.txt
pygame>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0